+++ src/test/java/lotto/model/CashierTest.java
@@ -1,6 +1,6 @@
 package lotto.model;
 
-import static lotto.LottoConstants.Price.LOTTO_PRICE_UNIT;
+import static lotto.rule.LottoConstants.Price.LOTTO_PRICE_UNIT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
